+++ BioFVM/BioFVM_basic_agent.cpp
@@ -76,12 +76,12 @@ Basic_Agent::Basic_Agent()
 	// register_microenvironment( default_microenvironment ); 
 
 	internalized_substrates = new std::vector<double>(0); // 
-	released_fractions_at_death = new std::vector<double>(0); 
+	fraction_released_at_death = new std::vector<double>(0); 
+	fraction_transferred_when_ingested = new std::vector<double>(0); 
 	register_microenvironment( get_default_microenvironment() );
 	
 	// these are done in register_microenvironment
 	// internalized_substrates.assign( get_default_microenvironment()->number_of_densities() , 0.0 ); 
-	// use_internal_densities_as_targets = false;  	
 	
 	return;	
 }
@@ -131,12 +131,14 @@ void Basic_Agent::set_internal_uptake_constants( double dt )
 	//   p(n+1) = (  p(n) + temp1 )/temp2
 	//int nearest_voxel= current_voxel_index;
 	
+/*	
 	// new for tracking internal densities
 	if( use_internal_densities_as_targets == true )
 	{
 		*saturation_densities = *internalized_substrates;
 		*saturation_densities /= ( 1e-15 + volume ); 
 	}
+*/
 	
 	double internal_constant_to_discretize_the_delta_approximation = dt * volume / ( (microenvironment->voxels(current_voxel_index)).volume ) ; // needs a fix 
 	
@@ -173,8 +175,8 @@ void Basic_Agent::register_microenvironment( Microenvironment* microenvironment_
 	internalized_substrates->resize( microenvironment->density_vector(0).size() , 0.0 );
 	total_extracellular_substrate_change.resize( microenvironment->density_vector(0).size() , 1.0 );
 	
-	released_fractions_at_death->resize( microenvironment->density_vector(0).size() , 0.0 ); 
-	use_internal_densities_as_targets = false;  
+	fraction_released_at_death->resize( microenvironment->density_vector(0).size() , 0.0 ); 
+	fraction_transferred_when_ingested->resize( microenvironment->density_vector(0).size() , 0.0 ); 
 
 	return; 
 }
@@ -186,13 +188,18 @@ Basic_Agent::~Basic_Agent()
 {
 	Microenvironment* pS = get_default_microenvironment(); 
 	
-	std::cout << "\t\t\t" << (*pS)(current_voxel_index) << "\t\t\t" << std::endl; 
+	// change in total in voxel: 
+	// total_ext = total_ext + fraction*total_internal 
+	// total_ext / vol_voxel = total_ext / vol_voxel + fraction*total_internal / vol_voxel 
+	// density_ext += fraction * total_internal / vol_volume 
+	
+	// std::cout << "\t\t\t" << (*pS)(current_voxel_index) << "\t\t\t" << std::endl; 
 	*internalized_substrates /=  pS->voxels(current_voxel_index).volume; // turn to density 
-	std::cout << "\t\t" << *internalized_substrates << std::endl; 
-	*internalized_substrates *= *released_fractions_at_death;  // what fraction is released? 
-	std::cout << "\t\t" << *internalized_substrates  << std::endl ;
+	*internalized_substrates *= *fraction_released_at_death;  // what fraction is released? 
+	
+	// release this amount into the environment 
+	
 	(*pS)(current_voxel_index) += *internalized_substrates; 
-	std::cout << "\t\t\t" << (*pS)(current_voxel_index) << "\t\t\t" << std::endl << std::endl; 
 	
 	return; 
 }