+++ custom_modules/custom.cpp
@@ -65,21 +65,11 @@
 ###############################################################################
 */
 
-#include "./custom-unit-substrate-conservation.h"
+#include "./custom.h"
 
 // declare cell definitions here 
 
-Cell_Definition motile_cell; 
-
-void inert_phenotype( Cell* pCell, Phenotype& phenotype , double dt )
-{
-	phenotype.cycle.data.transition_rate(0,1) = 0.0; 
-	phenotype.death.rates[0] = 0;
-	phenotype.death.rates[1] = 0;
-	
-	
-	return; 
-}
+Cell_Definition epithelial_cell , macrophage; 
 
 void create_cell_types( void )
 {
@@ -94,16 +84,11 @@ void create_cell_types( void )
 	
 	initialize_default_cell_definition();
 	cell_defaults.phenotype.secretion.sync_to_microenvironment( &microenvironment ); 
-
-	
-	// enable internalized substrate tracking 
-	
-	
 	
 	// Name the default cell type 
 	
 	cell_defaults.type = 0; 
-	cell_defaults.name = "tumor cell"; 
+	cell_defaults.name = "default"; 
 	
 	// set default cell cycle model 
 
@@ -135,39 +120,18 @@ void create_cell_types( void )
 
 	int G0G1_index = flow_cytometry_separated_cycle_model.find_phase_index( PhysiCell_constants::G0G1_phase );
 	int S_index = flow_cytometry_separated_cycle_model.find_phase_index( PhysiCell_constants::S_phase );
-	
-	int G2_index = flow_cytometry_separated_cycle_model.find_phase_index( PhysiCell_constants::G2_phase ); 
-	int M_index = flow_cytometry_separated_cycle_model.find_phase_index( PhysiCell_constants::M_phase ); 
 
 	// initially no necrosis 
 	cell_defaults.phenotype.death.rates[necrosis_model_index] = 0.0; 
-	
-	// set apoptosis rate 
-	cell_defaults.phenotype.death.rates[apoptosis_model_index] =  0.0025; 
 
-	// set a faster time scale for division 
-	cell_defaults.phenotype.cycle.data.transition_rate(G0G1_index,S_index) = 1.0 / 1.0; // 1 min wait time 
-	// let's speed up M phase 
-	cell_defaults.phenotype.cycle.data.transition_rate(M_index,G0G1_index) = 1.0 / 30.0; // 30 minutes M
-	// let's speed up G2 phase 
-	cell_defaults.phenotype.cycle.data.transition_rate(G2_index,M_index) = 1.0 / 60.0; // 60 minutes G2
-	// let's speed up S phase 
-	cell_defaults.phenotype.cycle.data.transition_rate(S_index,G2_index) = 1.0 / 90.0; // 90 minutes S
-	
 	// set oxygen uptake / secretion parameters for the default cell type 
 	cell_defaults.phenotype.secretion.uptake_rates[oxygen_substrate_index] = 10; 
 	cell_defaults.phenotype.secretion.secretion_rates[oxygen_substrate_index] = 0; 
 	cell_defaults.phenotype.secretion.saturation_densities[oxygen_substrate_index] = 38; 
-
-	// next substrate is secreted 
-	cell_defaults.phenotype.secretion.uptake_rates[1] = 0; 
-	cell_defaults.phenotype.secretion.secretion_rates[1] = 10; 
-	cell_defaults.phenotype.secretion.saturation_densities[1] = 1; 
-
-	// next substrate is secreted and uptake 
-	cell_defaults.phenotype.secretion.uptake_rates[2] = 3; 
-	cell_defaults.phenotype.secretion.secretion_rates[2] = 10; 
-	cell_defaults.phenotype.secretion.saturation_densities[2] = 1; 
+	
+	// add custom data here, if any 
+	
+	
 	
 
 	// Now, let's define another cell type. 
@@ -176,29 +140,33 @@ void create_cell_types( void )
 	// make this cell type randomly motile, less adhesive, greater survival, 
 	// and less proliferative 
 	
-	motile_cell = cell_defaults; 
-	motile_cell.type = 1; 
-	motile_cell.name = "motile tumor cell"; 
+	epithelial_cell.type = 1; 
+	epithelial_cell.name = "epithelial cell"; 
 	
 	// make sure the new cell type has its own reference phenotype
 	
-	motile_cell.parameters.pReference_live_phenotype = &( motile_cell.phenotype ); 
+	epithelial_cell.parameters.pReference_live_phenotype = &( epithelial_cell.phenotype ); 
+	
+	// epithelial cells are not random 
+	
+	
+	
 	
 	// enable random motility 
-	motile_cell.phenotype.motility.is_motile = true; 
-	motile_cell.phenotype.motility.persistence_time = parameters.doubles( "motile_cell_persistence_time" ); // 15.0; 
-	motile_cell.phenotype.motility.migration_speed = parameters.doubles( "motile_cell_migration_speed" ); // 0.25 micron/minute 
-	motile_cell.phenotype.motility.migration_bias = 0.0;// completely random 
+	epithelial_cell.phenotype.motility.is_motile = true; 
+	epithelial_cell.phenotype.motility.persistence_time = parameters.doubles( "motile_cell_persistence_time" ); // 15.0; 
+	epithelial_cell.phenotype.motility.migration_speed = parameters.doubles( "motile_cell_migration_speed" ); // 0.25 micron/minute 
+	epithelial_cell.phenotype.motility.migration_bias = 0.0;// completely random 
 	
 	// Set cell-cell adhesion to 5% of other cells 
-	motile_cell.phenotype.mechanics.cell_cell_adhesion_strength *= parameters.doubles( "motile_cell_relative_adhesion" ); // 0.05; 
+	epithelial_cell.phenotype.mechanics.cell_cell_adhesion_strength *= parameters.doubles( "motile_cell_relative_adhesion" ); // 0.05; 
 	
 	// Set apoptosis to zero 
-	motile_cell.phenotype.death.rates[apoptosis_model_index] = parameters.doubles( "motile_cell_apoptosis_rate" ); // 0.0; 
+	epithelial_cell.phenotype.death.rates[apoptosis_model_index] = parameters.doubles( "motile_cell_apoptosis_rate" ); // 0.0; 
 	
 	// Set proliferation to 10% of other cells. 
 	// Alter the transition rate from G0G1 state to S state
-	motile_cell.phenotype.cycle.data.transition_rate(G0G1_index,S_index) *= 
+	epithelial_cell.phenotype.cycle.data.transition_rate(G0G1_index,S_index) *= 
 		parameters.doubles( "motile_cell_relative_cycle_entry_rate" ); // 0.1; 
 	
 	return; 
@@ -224,46 +192,27 @@ void setup_microenvironment( void )
 
 	default_microenvironment_options.calculate_gradients = false; 
 	
-	// set Dirichlet conditions off
+	// do not use oxygen as first substrate 
+	default_microenvironment_options.use_oxygen_as_first_field = false; 
+	
+	// set properties 
+	microenvironment.set_density( 0 , "viral particle density", "particles/micron^3" ); 
+	microenvironment.diffusion_coefficients[0] = 1e4; 
+	microenvironment.decay_rates[0] = 0; 
+	
+	// set Dirichlet conditions 
 
 	default_microenvironment_options.outer_Dirichlet_conditions = false;
-
-	// enable tracking (requires 1.5.0 or later)
 	
-	default_microenvironment_options.track_internalized_substrates_in_each_agent = true; 
-	
-	
-	// add a new substrate 
-	
-	microenvironment.add_density( "sample 1", "dimensionless" ); 
-	microenvironment.diffusion_coefficients[1] = 1e3; 
-	microenvironment.decay_rates[1] = 0; 	
-	
-	// add the immunostimulatory factor 
-	
-	microenvironment.add_density( "sample 2", "dimensionless" ); 
-	microenvironment.diffusion_coefficients[2] = 2e3; 
-	microenvironment.decay_rates[2] = 0; 	
-	
-	// use this for default Dirichlet conditions. 
-	// Actually, we'll just use it for initial conditions 
-	std::vector<double> bc_vector( 3 , 160  ); // 21% o2
-	bc_vector[1] = 0.5; 
-	bc_vector[2] = 0.5; 
+	// if there are more substrates, resize accordingly 
+	std::vector<double> bc_vector( 1 , 0.0 ); // 5% o2
 	default_microenvironment_options.Dirichlet_condition_vector = bc_vector;
+	
+	// 
 	
 	// initialize BioFVM 
 	
 	initialize_microenvironment(); 	
-	
-	// set the oxygen decay rate back to zero 
-	microenvironment.decay_rates[0] = 0; 	
-		
-	// now, let's set all the substrates to the bc_vector value 
-	for( unsigned int n=0; n < microenvironment.number_of_voxels() ; n++ )
-	{
-		microenvironment(n) = bc_vector; 
-	}	
 	
 	return; 
 }
@@ -271,37 +220,22 @@ void setup_microenvironment( void )
 void setup_tissue( void )
 {
 	// create some cells near the origin
-	std::cout << __FILE__ << " " << __LINE__ << std::endl; 
+	
 	Cell* pC;
+
+	pC = create_cell(); 
+	pC->assign_position( 0.0, 0.0, 0.0 );
+
+	pC = create_cell(); 
+	pC->assign_position( -100, 0, 0.0 );
 	
-	double theta = 0.0; 
-	double r = 20; 
+	pC = create_cell(); 
+	pC->assign_position( 0, 100, 0.0 );
 	
-	double dr = 1.25; 
-	double dS = 20; 
+	// now create a motile cell 
 	
-	double temp = sqrt( dS*dS - dr*dr ); 
-
-	std::cout << __FILE__ << " " << __LINE__ << std::endl; 
-
-	int G0G1_index = flow_cytometry_separated_cycle_model.find_phase_index( PhysiCell_constants::G0G1_phase );
-	int S_index = flow_cytometry_separated_cycle_model.find_phase_index( PhysiCell_constants::S_phase );
-
-	std::cout << __FILE__ << " " << __LINE__ << std::endl; 
-
-	int i = 0; 
-	while( r < 425 )
-	{
-		std::cout << i << " " ; 
-		pC = create_cell(); 
-		pC->assign_position( r*cos(theta), r*sin(theta), 0.0 ); 
-
-		theta += temp / r; 
-		r += dr; 
-
-		i++;		
-	}
-	std::cout << __FILE__ << " " << __LINE__ << std::endl; 
+	pC = create_cell( motile_cell ); 
+	pC->assign_position( 15.0, -18.0, 0.0 );
 	
 	return; 
 }
@@ -319,26 +253,4 @@ std::vector<std::string> my_coloring_function( Cell* pCell )
 	}
 	
 	return output; 
-}
-
-std::vector<double> integrate_total_substrates( void )
-{
-	// start with 0 vector 
-	std::vector<double> out( microenvironment.number_of_densities() , 0.0 ); 
-
-	// integrate extracellular substrates 
-	for( unsigned int n = 0; n < microenvironment.number_of_voxels() ; n++ )
-	{
-		// out = out + microenvironment(n) * dV(n) 
-		axpy( &out , microenvironment.mesh.voxels[n].volume , microenvironment(n) ); 
-	}
-
-	// inte
-	for( unsigned int n=0; n < (*all_cells).size(); n++ )
-	{
-		Cell* pC = (*all_cells)[n];
-		out += *(pC->internalized_substrates);
-	}
-	
-	return out; 
 }