+++ main.cpp
@@ -78,7 +78,7 @@
 
 // put custom code modules here! 
 
-#include "./custom_modules/custom-unit-substrate-conservation.h" 
+#include "./custom_modules/custom.h" 
 	
 using namespace BioFVM;
 using namespace PhysiCell;
@@ -91,7 +91,7 @@ int main( int argc, char* argv[] )
 	if( argc > 1 )
 	{ XML_status = load_PhysiCell_config_file( argv[1] ); }
 	else
-	{ XML_status = load_PhysiCell_config_file( "./config/unit-conservation-settings.xml" ); }
+	{ XML_status = load_PhysiCell_config_file( "./config/PhysiCell_settings.xml" ); }
 	if( !XML_status )
 	{ exit(-1); }
 	
@@ -100,8 +100,6 @@ int main( int argc, char* argv[] )
 	
 	// time setup 
 	std::string time_units = "min"; 
-
-		std::cout << __LINE__ << std::endl; 
 
 	/* Microenvironment setup */ 
 	
@@ -112,14 +110,12 @@ int main( int argc, char* argv[] )
 	// set mechanics voxel size, and match the data structure to BioFVM
 	double mechanics_voxel_size = 30; 
 	Cell_Container* cell_container = create_cell_container_for_microenvironment( microenvironment, mechanics_voxel_size );
-	std::cout << __LINE__ << std::endl; 
 	
 	/* Users typically start modifying here. START USERMODS */ 
 	
 	create_cell_types();
-	std::cout << __LINE__ << std::endl; 
+	
 	setup_tissue();
-	std::cout << __LINE__ << std::endl; 
 
 	/* Users typically stop modifying here. END USERMODS */ 
 	
@@ -130,9 +126,6 @@ int main( int argc, char* argv[] )
 	set_save_biofvm_cell_data( true ); 
 	set_save_biofvm_cell_data_as_custom_matlab( true );
 	
-	std::cout << __LINE__ << std::endl; 
-	
-
 	// save a simulation snapshot 
 	
 	char filename[1024];
@@ -147,7 +140,7 @@ int main( int argc, char* argv[] )
 	// for simplicity, set a pathology coloring function 
 	
 	std::vector<std::string> (*cell_coloring_function)(Cell*) = my_coloring_function; 
-
+	
 	sprintf( filename , "%s/initial.svg" , PhysiCell_settings.folder.c_str() ); 
 	SVG_plot( filename , microenvironment, 0.0 , PhysiCell_globals.current_time, cell_coloring_function );
 	
@@ -167,40 +160,7 @@ int main( int argc, char* argv[] )
 		report_file<<"simulated time\tnum cells\tnum division\tnum death\twall time"<<std::endl;
 	}
 	
-	//
-	std::cout << __LINE__ << std::endl; 
-	Cell* pCell_1 = create_cell(); //  cell_defaults ); 
-	pCell_1->assign_position( 0,0,0 ); 
-	std::cout << __LINE__ << std::endl; 
-	Cell* pCell_2 = pCell_1->divide(); 
-	std::cout << __LINE__ << std::endl; 
-	
-	std::cout << "1:" << pCell_1->phenotype.secretion.secretion_rates << std::endl; 
-	std::cout << "2:" << pCell_2->phenotype.secretion.secretion_rates << std::endl; 
-	
-	// change cell 1 secretion rate 
-	pCell_1->phenotype.secretion.secretion_rates[0] = 12; 
-	
-	std::cout << "1:" << pCell_1->phenotype.secretion.secretion_rates << std::endl; 
-	std::cout << "2:" << pCell_2->phenotype.secretion.secretion_rates << std::endl; 
-
-	system( "pause" ); 
-	
-	Bools bools; 
-	for( int i=0 ; i < bools.size() ; i++ )
-	{
-		std::cout << i << " : " << bools.values[i] << std::endl; 
-	}
-	
-	// advance_molecular_models( molecular_dt ); 
-	
-
-	system("pause"); 
-	
 	// main loop 
-	
-	std::cout << "Unit test: conservation with individual agent substrate internalization " << std::endl 
-		<< "If this works, the total amount of each substrate should stay fixed at each output " << std::endl << std::endl ; 
 	
 	try 
 	{		
@@ -237,49 +197,17 @@ int main( int argc, char* argv[] )
 					PhysiCell_globals.SVG_output_index++; 
 					PhysiCell_globals.next_SVG_save_time  += PhysiCell_settings.SVG_save_interval;
 				}
-				
-				std::cout << "Total substrates " << integrate_total_substrates() << std::endl; 
 			}
 
 			// update the microenvironment
 			microenvironment.simulate_diffusion_decay( diffusion_dt );
 			
-			// run molecular models 
-			advance_molecular_models( molecular_dt ); 
-
 			// run PhysiCell 
 			((Cell_Container *)microenvironment.agent_container)->update_all_cells( PhysiCell_globals.current_time );
 			
 			/*
 			  Custom add-ons could potentially go here. 
 			*/
-			
-			static double next_time = 60.0; 
-			if( fabs( PhysiCell_globals.current_time - next_time ) < 1e-6 && 
-				PhysiCell_globals.current_time < 1440.1 && 
-				(*all_cells).size() > 2 )
-			{
-				// choose a cell to eat the "next" cell 
-				int n = (int) round( UniformRandom() * ((*all_cells).size()-2) ); 
-				
-				Cell* pC1 = (*all_cells)[n];
-				Cell* pC2 = (*all_cells)[n+1]; 
-				
-				std::cout << std::endl << std::endl << "\tCell " << pC1 << " is ingesting cell " << pC2
-					<< std::endl << std::endl; 
-				
-				// ingest the cell 
-				pC1->ingest_cell( pC2 ); 
-				
-				// set the "eater" to zero division and death rates 
-				pC1->functions.update_phenotype = inert_phenotype; 
-				
-				// if we accidentally had a zombie cell do the eating, kill it off. 
-				if( pC1->phenotype.death.dead == true )
-				{ pC1->die(); } 
-	
-				next_time += 90.0; 			
-			}
 			
 			PhysiCell_globals.current_time += diffusion_dt;
 		}